5.8-p023 parser toolkits
Source code for LEF/DEF version 5.8 P023 parser toolkit including C++
and C APIs to the parser includes the following changes:New callbacks
for the LEF parser: extern void
lefrSetMacroSiteCbk(lefrMacroSiteCbkFnType); extern void
lefrSetMacroForeignCbk(lefrMacroForeignCbkFnType); extern void
lefrUnsetMacroSiteCbk(); extern void lefrUnsetMacroForeignCbk(); The
parsers contained in the Licensed Materials are required to compile the
LEF and DEF translator code supplied with the Open Access 22.50 release.
They should be used when compiling the translators in OA 22.50-P064 and
later releases. The kits contained in the Licensed Materials are not
compatible with the Open Access 22.43 release. The parsers supplied in
the lef/def_5.8-p002.tar.Z kits should be used to compile the lef/def
translators supplied with the OpenAccess 22.43 release.